requirements.txt
# Core RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.3.7

# Vector database
chromadb>=0.4.18

# Document processing
pypdf2>=3.0.1
python-docx>=1.1.0
# For encrypted PDFs
pycryptodome>=3.15.0

# Web scraping
beautifulsoup4>=4.12.2
requests>=2.31.0
selenium>=4.15.2
webdriver-manager>=4.0.1

# Text processing
tiktoken>=0.5.1
sentence-transformers>=2.2.2

# Vector search
faiss-cpu>=1.7.4

# Data processing (flexible versions)
numpy>=1.21.0
pandas>=2.0.0

# Web interface
streamlit>=1.28.0

# Environment and utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
aiohttp>=3.9.0

# Local LLM dependencies (optional)
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
optimum>=1.16.0

# Ollama support (optional)
ollama>=0.1.7

# Alternative embedding models (nomic-ai alternative)
# nomic-ai is not available for all Python versions
# Using sentence-transformers as the default embedding model 